Split module and storage
