+++ src/many-ledger/src/storage/ledger.rs
@@ -1,0 +1,44 @@
+use crate::storage::{key_for_account_balance, LedgerStorage};
+use many_identity::Address;
+use many_types::ledger::{Symbol, TokenAmount};
+use std::collections::{BTreeMap, BTreeSet};
+
+impl LedgerStorage {
+    fn get_all_balances(&self, identity: &Address) -> BTreeMap<&Symbol, TokenAmount> {
+        if identity.is_anonymous() {
+            // Anonymous cannot hold funds.
+            BTreeMap::new()
+        } else {
+            let mut result = BTreeMap::new();
+            for symbol in self.symbols.keys() {
+                match self
+                    .persistent_store
+                    .get(&key_for_account_balance(identity, symbol))
+                {
+                    Ok(None) => {}
+                    Ok(Some(value)) => {
+                        result.insert(symbol, TokenAmount::from(value));
+                    }
+                    Err(_) => {}
+                }
+            }
+
+            result
+        }
+    }
+
+    pub fn get_multiple_balances(
+        &self,
+        identity: &Address,
+        symbols: &BTreeSet<Symbol>,
+    ) -> BTreeMap<&Symbol, TokenAmount> {
+        if symbols.is_empty() {
+            self.get_all_balances(identity)
+        } else {
+            self.get_all_balances(identity)
+                .into_iter()
+                .filter(|(k, _v)| symbols.contains(*k))
+                .collect()
+        }
+    }
+}
